refactor(core): move DIR_TO_TYPE and TYPE_TO_DIR to common.types.ts
Canonical record-type ↔ directory mappings now live alongside
GitGovRecordType and are exported via the Records namespace.
id_parser.ts re-exports for internal backwards compatibility.
